requirements.txt
-f /usr/share/pip-wheels
asgiref==3.3.1
Django==3.1.4
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-cors-headers==3.6.0
django-filter==2.4.0
django-registration-redux==2.9
djangorestframework==3.12.2
importlib-metadata==3.4.0
Markdown==3.3.3
mysqlclient==2.0.3
python-dotenv==0.16.0
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
